Smooth lesson loading transitions, show Jiki during auth check
Keep a single LessonLoadingModal instance mounted across page fetch,
dynamic chunk load, and exercise loader stages so its CSS animations
don't restart at each handoff. Replace the auth-gate spinner with
LoadingJiki so the wakeup image actually shows on hard refresh.
