+++ app/components/lesson/Lesson.tsx
@@ -1,0 +1,90 @@
+"use client";
+
+import LessonLoadingModal from "@/components/common/LessonLoadingModal/LessonLoadingModal";
+import { fetchUserCourse } from "@/lib/api/courses";
+import { fetchLesson, fetchUserLesson } from "@/lib/api/lessons";
+import type { UserCourse } from "@/types/course";
+import type { LessonWithData } from "@/types/lesson";
+import { useCallback, useEffect, useState } from "react";
+import LessonContent from "./LessonContent";
+import LessonError from "./LessonError";
+
+interface LessonProps {
+  slug: string;
+}
+
+export default function Lesson({ slug }: LessonProps) {
+  const [lesson, setLesson] = useState<LessonWithData | null>(null);
+  const [userCourse, setUserCourse] = useState<UserCourse | null>(null);
+  const [isCompleted, setIsCompleted] = useState(false);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+  const [innerReady, setInnerReady] = useState(false);
+
+  const handleReady = useCallback(() => setInnerReady(true), []);
+
+  // Update document title when lesson loads
+  useEffect(() => {
+    if (lesson) {
+      document.title = `${lesson.title} - Jiki`;
+    }
+  }, [lesson]);
+
+  // Load lesson and user course on mount
+  useEffect(() => {
+    let cancelled = false;
+
+    async function loadData() {
+      try {
+        setLoading(true);
+        const [lessonData, userCourseData, userLessonResult] = await Promise.all([
+          fetchLesson(slug),
+          fetchUserCourse(),
+          fetchUserLesson(slug).catch(() => null)
+        ]);
+        if (cancelled) {
+          return;
+        }
+        setLesson(lessonData);
+        setUserCourse(userCourseData);
+        setIsCompleted(userLessonResult?.status === "completed");
+      } catch (err) {
+        if (!cancelled) {
+          console.error("Failed to fetch lesson:", err);
+          // Auth/network/rate-limit errors never reach here (handled globally)
+          // Only application errors (404, 500, validation) reach this catch block
+          setError(err instanceof Error ? err.message : "Failed to load lesson");
+        }
+      } finally {
+        if (!cancelled) {
+          setLoading(false);
+        }
+      }
+    }
+
+    void loadData();
+
+    return () => {
+      cancelled = true;
+    };
+  }, [slug]);
+
+  if (error) {
+    return <LessonError error={error} />;
+  }
+
+  const showModal = loading || !lesson || !innerReady;
+
+  // LessonContent must mount *underneath* the modal (not behind an early return) so its
+  // dynamic chunk and exercise loader can run in the background. The child fires onReady
+  // when truly ready, which flips innerReady and unmounts the modal in a single render —
+  // keeping one modal instance alive across the whole load so its CSS animations don't restart.
+  return (
+    <>
+      {lesson && (
+        <LessonContent lesson={lesson} userCourse={userCourse} isCompleted={isCompleted} onReady={handleReady} />
+      )}
+      {showModal && <LessonLoadingModal />}
+    </>
+  );
+}
